fix(aux): 辅助块的身份改用包序号，绝对文件偏移在管道输入下不成立
管道读入解压全盘失败：先验被登记在 @0(不可 seek 时 getCurrentBlockStart 退化为 0)，
数据块却按包内相对偏移 235 查表，全部落空，解码侧的致命路径把每个块都判死。

根因是我选错了辅助块的身份。绝对文件偏移只在可 seek 的文件上成立，
管道下根本没有这个概念。包序号只依赖已解析的包头个数，文件/管道/cat 一致成立。
块 meta 里的偏移保留，但降级为 seek 手段与校验值，不再承担索引职责。

当前假定一个包里每种辅助块至多一份；分片落地后键需扩成 (包序号, 包内序号)，
已在 AuxBlockConsumer::claim 的注释里写明。

七条读路径全部验证：管道读入/单包带参考/单包无参考/流式/区域查询/解压到管道/gz。

+++ src/compress_engine.h
@@ -146,7 +146,7 @@ public:
         return qualPriorOffset.load(std::memory_order_acquire);
     }
 
-    AuxPayloadPtr getQualPrior(int64_t /*blockAddress*/) override {
+    AuxPayloadPtr getQualPrior(int32_t /*packageIndex*/) override {
         /* 在串行首块窗口内写入，随后只读；地址原子量的 acquire 同时为它建立可见性。 */
         return (getQualPriorAddress() < 0) ? AuxPayloadPtr() : qualPriorBlob;
     }
